update: unified parser and method descriptions, also button colors updated
