+++ app/src/main/java/org/satochip/seedkeeper/ui/components/home/SecretsFilter.kt
@@ -2,7 +2,6 @@ package org.satochip.seedkeeper.ui.components.home
 
 import androidx.compose.foundation.Image
 import androidx.compose.foundation.background
-import androidx.compose.foundation.clickable
 import androidx.compose.foundation.layout.Column
 import androidx.compose.foundation.layout.Row
 import androidx.compose.foundation.layout.size
