+++ app/src/main/java/org/satochip/seedkeeper/ui/views/generate/GeneratePassword.kt
@@ -1,31 +1,26 @@
 package org.satochip.seedkeeper.ui.views.generate
 
 import android.content.SharedPreferences
-import androidx.compose.foundation.layout.Arrangement
 import androidx.compose.foundation.layout.Column
-import androidx.compose.foundation.layout.Row
 import androidx.compose.foundation.layout.Spacer
-import androidx.compose.foundation.layout.fillMaxSize
 import androidx.compose.foundation.layout.fillMaxWidth
 import androidx.compose.foundation.layout.height
 import androidx.compose.runtime.Composable
 import androidx.compose.runtime.MutableState
 import androidx.compose.ui.Alignment
 import androidx.compose.ui.Modifier
-import androidx.compose.ui.graphics.Color
 import androidx.compose.ui.unit.dp
 import org.satochip.client.seedkeeper.SeedkeeperSecretType
 import org.satochip.seedkeeper.R
-import org.satochip.seedkeeper.data.SecretData
 import org.satochip.seedkeeper.data.GenerateViewItems
 import org.satochip.seedkeeper.data.PasswordOptions
+import org.satochip.seedkeeper.data.SecretData
 import org.satochip.seedkeeper.data.SeedkeeperPreferences
+import org.satochip.seedkeeper.ui.components.generate.ButtonsField
 import org.satochip.seedkeeper.ui.components.generate.InputField
 import org.satochip.seedkeeper.ui.components.generate.PasswordLengthField
 import org.satochip.seedkeeper.ui.components.generate.SecretTextField
-import org.satochip.seedkeeper.ui.components.shared.SatoButton
 import org.satochip.seedkeeper.ui.components.shared.TitleTextField
-import org.satochip.seedkeeper.ui.theme.SatoActiveTracer
 import org.satochip.seedkeeper.ui.theme.SatoPurple
 import org.satochip.seedkeeper.utils.isClickable
 
@@ -91,57 +86,38 @@ fun GeneratePassword(
                 onClick(GenerateViewItems.COPY_TO_CLIPBOARD, secret.value, null)
             }
         )
-        Row(
-            modifier = Modifier.fillMaxSize(),
-            horizontalArrangement = Arrangement.Center
-        ) {
-            //Generate
-            SatoButton(
-                modifier = Modifier
-                    .weight(1f),
-                onClick = {
-                    secret.value = onClick(
-                        GenerateViewItems.GENERATE_A_PASSWORD,
-                        null,
-                        passwordOptions.value
-                    )
-                },
-                text = if(secret.value.isNotEmpty()) R.string.regenerate else R.string.generate,
-            )
-            //Import
-            SatoButton(
-                modifier = Modifier
-                    .weight(1f),
-                onClick = {
-                    if (isClickable(secret, curValueLabel)) {
-                        onImportSecret(
-                            SecretData(
-                                size = passwordOptions.value.passwordLength,
-                                type = SeedkeeperSecretType.PASSWORD,
-                                password = secret.value,
-                                label = curValueLabel.value,
-                                login = curValueLogin.value,
-                                url = curValueUrl.value,
-                            )
+        ButtonsField(
+            secret = secret,
+            curValueLabel = curValueLabel,
+            onGenerateClick = {
+                secret.value = onClick(
+                    GenerateViewItems.GENERATE_A_PASSWORD,
+                    null,
+                    passwordOptions.value
+                )
+            },
+            onImportSecret = {
+                if (isClickable(secret, curValueLabel)) {
+                    onImportSecret(
+                        SecretData(
+                            size = passwordOptions.value.passwordLength,
+                            type = SeedkeeperSecretType.PASSWORD,
+                            password = secret.value,
+                            label = curValueLabel.value,
+                            login = curValueLogin.value,
+                            url = curValueUrl.value,
                         )
-                    }
-                    if (curValueLogin.value.isNotEmpty()) {
-                        val stringSet = listOf(curValueLogin.value).toSet()
-                        retrievedSet.value += stringSet
-                        settings.edit().putStringSet(
-                            SeedkeeperPreferences.USED_LOGINS.name,
-                            retrievedSet.value
-                        ).apply()
-                    }
-                },
-                text = R.string.importButton,
-                textColor = if (
-                    isClickable(
-                        secret,
-                        curValueLabel
                     )
-                ) Color.White else SatoActiveTracer
-            )
-        }
+                }
+                if (curValueLogin.value.isNotEmpty()) {
+                    val stringSet = listOf(curValueLogin.value).toSet()
+                    retrievedSet.value += stringSet
+                    settings.edit().putStringSet(
+                        SeedkeeperPreferences.USED_LOGINS.name,
+                        retrievedSet.value
+                    ).apply()
+                }
+            }
+        )
     }
 }
