+++ app/src/main/java/org/satochip/seedkeeper/ui/views/import/ImportMnemonic.kt
@@ -28,7 +28,8 @@ import org.satochip.seedkeeper.ui.components.generate.SecretTextField
 import org.satochip.seedkeeper.ui.components.import.MnemonicImportField
 import org.satochip.seedkeeper.ui.components.shared.SatoButton
 import org.satochip.seedkeeper.ui.components.shared.TitleTextField
-import org.satochip.seedkeeper.ui.theme.SatoActiveTracer
+import org.satochip.seedkeeper.ui.theme.SatoButtonPurple
+import org.satochip.seedkeeper.ui.theme.SatoInactiveTracer
 import org.satochip.seedkeeper.ui.theme.SatoPurple
 import org.satochip.seedkeeper.utils.isClickable
 
@@ -123,12 +124,12 @@ fun ImportMnemonic(
                     }
                 },
                 text = R.string.importButton,
-                textColor = if (
+                buttonColor = if (
                     isClickable(
                         secret,
                         curValueLabel
                     )
-                ) Color.White else SatoActiveTracer
+                ) SatoButtonPurple else SatoButtonPurple.copy(alpha = 0.6f),
             )
         }
     }
