+++ app/src/main/java/org/satochip/seedkeeper/ui/views/import/ImportWalletDescriptor.kt
@@ -26,7 +26,7 @@ import org.satochip.seedkeeper.ui.components.generate.InputField
 import org.satochip.seedkeeper.ui.components.generate.SecretTextField
 import org.satochip.seedkeeper.ui.components.shared.SatoButton
 import org.satochip.seedkeeper.ui.components.shared.TitleTextField
-import org.satochip.seedkeeper.ui.theme.SatoActiveTracer
+import org.satochip.seedkeeper.ui.theme.SatoButtonPurple
 import org.satochip.seedkeeper.ui.theme.SatoPurple
 import org.satochip.seedkeeper.utils.isClickable
 
@@ -100,12 +100,12 @@ fun ImportWalletDescriptor(
                     }
                 },
                 text = R.string.importButton,
-                textColor = if (
+                buttonColor = if (
                     isClickable(
                         secret,
                         curValueLabel
                     )
-                ) Color.White else SatoActiveTracer
+                ) SatoButtonPurple else SatoButtonPurple.copy(alpha = 0.6f),
             )
         }
     }
