+++ app/src/main/java/org/satochip/seedkeeper/ui/views/settings/SettingsView.kt
@@ -21,8 +21,7 @@ import org.satochip.seedkeeper.ui.components.settings.SatoDescriptionField
 import org.satochip.seedkeeper.ui.components.settings.SatoToggleButton
 import org.satochip.seedkeeper.ui.components.shared.HeaderAlternateRow
 import org.satochip.seedkeeper.ui.components.shared.SatoButton
-import org.satochip.seedkeeper.ui.theme.SatoButtonBlue
-import org.satochip.seedkeeper.ui.theme.SatoInactiveTracer
+import org.satochip.seedkeeper.ui.theme.SatoButtonPurple
 
 @Composable
 fun SettingsView(
@@ -91,7 +90,7 @@ fun SettingsView(
                     }
                 },
                 text = R.string.showLogs,
-                buttonColor = if (debugMode.value) SatoButtonBlue else SatoInactiveTracer,
+                buttonColor = if (debugMode.value) SatoButtonPurple else SatoButtonPurple.copy(0.6f),
             )
         }
     }
