+++ app/src/main/java/org/satochip/seedkeeper/utils/Utils.kt
@@ -1,236 +1,15 @@
 package org.satochip.seedkeeper.utils
 
-import android.util.Patterns
 import androidx.compose.runtime.MutableState
-import org.bitcoinj.crypto.MnemonicCode
-import org.satochip.client.seedkeeper.SeedkeeperSecretType
-import org.satochip.seedkeeper.data.SecretData
-import org.satochip.seedkeeper.services.SatoLog
-import java.nio.ByteBuffer
 
 const val TAG = "Utlis"
 
-//Generate view
 fun isClickable(
     secret: MutableState<String>,
     curValueLabel: MutableState<String>
 ): Boolean {
     return secret.value.isNotEmpty()
             && curValueLabel.value.isNotEmpty()
-}
-
-fun isEmailCorrect(
-    curValueLogin: MutableState<String>
-): Boolean {
-    return curValueLogin.value.isNotEmpty() && Patterns.EMAIL_ADDRESS.matcher(curValueLogin.value).matches()
-}
-
-fun stringToList(inputString: String?): List<String?>? {
-    return inputString?.split("\\s+".toRegex())
-}
-
-fun parseMasterseedMnemonicCardData(bytes: ByteArray): SecretData? {
-    var index = 0
-    var descriptor = ""
-
-    if (bytes.isEmpty()) {
-        SatoLog.e(TAG, "Byte array is empty")
-        return null
-    }
-    val masterseedSize = bytes[index].toUByte().toInt()
-    index++
-
-    if (masterseedSize < 0 || index + masterseedSize > bytes.size) {
-        SatoLog.e(TAG, "Invalid masterseedSize")
-        return null
-    }
-    val masterseedBytes = bytes.copyOfRange(index, index + masterseedSize) // todo: find usage
-
-    index += masterseedSize
-    index++
-    val entropySize = bytes[index].toUByte().toInt()
-    index++
-
-    if (entropySize < 0 || index + entropySize > bytes.size) {
-        SatoLog.e(TAG, "Invalid mnemonic size")
-        return null
-    }
-    val entropyBytes = bytes.copyOfRange(index, index + entropySize)
-
-    val mnemonic = MnemonicCode.INSTANCE.toMnemonic(entropyBytes).joinToString(separator = " ")
-    index += entropySize
-
-    var passphrase: String? = null
-    if (index < bytes.size) {
-        val passphraseSize = bytes[index].toUByte().toInt()
-        index++
-        if (passphraseSize > 0 && index + passphraseSize <= bytes.size) {
-            val passphraseBytes = bytes.copyOfRange(index, index + passphraseSize)
-            index += passphraseSize
-            passphrase = String(passphraseBytes, Charsets.UTF_8)
-        }
-    }
-    if (index < bytes.size) {
-        val descriptorSizeArray = bytes.sliceArray(index..index + 1)
-        val descriptorSize = ByteBuffer.wrap(descriptorSizeArray).short
-        index += 2
-        if (index + descriptorSize > bytes.size) {
-            SatoLog.e(TAG, "Invalid descriptor size")
-            return null
-        }
-        val descriptorBytes = bytes.copyOfRange(index, index + descriptorSize)
-        descriptor = String(descriptorBytes, Charsets.UTF_8)
-    }
-
-
-    return SecretData(
-        password = passphrase ?: "",
-        mnemonic = mnemonic,
-        size = countWords(mnemonic),
-        label = "",
-        type = SeedkeeperSecretType.MASTERSEED,
-        descriptor = descriptor
-    )
-}
-
-fun parseMnemonicCardData(bytes: ByteArray): SecretData? {
-    var index = 0
-
-    if (bytes.isEmpty()) {
-        SatoLog.e(TAG, "Byte array is empty")
-        return null
-    }
-    val mnemonicSize = bytes[index].toUByte().toInt()
-    index += 1
-    if (mnemonicSize < 0 || index + mnemonicSize > bytes.size) {
-        SatoLog.e(TAG, "Invalid mnemonic size")
-        return null
-    }
-    val mnemonicBytes = bytes.copyOfRange(index, index + mnemonicSize)
-    index += mnemonicSize
-    val mnemonic = String(mnemonicBytes, Charsets.UTF_8)
-    if (mnemonic.isEmpty()) {
-        SatoLog.e(TAG, "Mnemonic bytes conversion to string failed")
-        return null
-    }
-
-    var passphrase: String? = null
-    if (index < bytes.size) {
-        val passphraseSize = bytes[index].toUByte().toInt()
-        index += 1
-        if (passphraseSize > 0 && index + passphraseSize <= bytes.size) {
-            val passphraseBytes = bytes.copyOfRange(index, index + passphraseSize)
-            index += passphraseSize
-            passphrase = String(passphraseBytes, Charsets.UTF_8)
-        }
-    }
-
-    return SecretData(
-        password = passphrase ?: "",
-        mnemonic = mnemonic,
-        size = countWords(mnemonic),
-        label = "",
-        type = SeedkeeperSecretType.BIP39_MNEMONIC
-    )
-}
-
-fun parsePubkeyData(bytes: ByteArray): SecretData? {
-    val pubkeySize = bytes[0].toInt()
-    if (1 + pubkeySize > bytes.size) {
-        SatoLog.e(TAG, "Invalid pubkey size")
-        return null
-    }
-    val pubkeyBytes = bytes.copyOfRange(1, 1 + pubkeySize)
-    val hexString = pubkeyBytes.joinToString(separator = "") { byte -> "%02x".format(byte) }
-
-    return SecretData(
-        password = hexString,
-        login = "",
-        url = "",
-        label = "",
-        type = SeedkeeperSecretType.PUBKEY,
-        size = 0,
-    )
-}
-
-fun parseWalletDescriptorData(bytes: ByteArray): SecretData? {
-    var index = 0
-
-    val descriptorSizeArray = bytes.sliceArray(0..1)
-    val descriptorSize = ByteBuffer.wrap(descriptorSizeArray).short
-    index += 2
-    if (index + descriptorSize > bytes.size) {
-        SatoLog.e(TAG, "Invalid descriptor size")
-        return null
-    }
-    val descriptorBytes = bytes.copyOfRange(index, index + descriptorSize)
-    val descriptor = String(descriptorBytes, Charsets.UTF_8)
-    if (descriptor.isEmpty()) {
-        SatoLog.e(TAG, "Descriptor bytes conversion to string failed")
-        return null
-    }
-    return SecretData(
-        password = "",
-        login = "",
-        url = "",
-        label = "",
-        type = SeedkeeperSecretType.DATA,
-        size = 0,
-        descriptor = descriptor
-    )
-}
-
-fun parsePasswordCardData(bytes: ByteArray): SecretData? {
-    var index = 0
-
-    val passwordSize = bytes[index].toInt()
-    index += 1
-    if (index + passwordSize > bytes.size) {
-        SatoLog.e(TAG, "Invalid password size")
-        return null
-    }
-    val passwordBytes = bytes.copyOfRange(index, index + passwordSize)
-    index += passwordSize
-    val password = String(passwordBytes, Charsets.UTF_8)
-    if (password.isEmpty()) {
-        SatoLog.e(TAG, "Password bytes conversion to string failed")
-        return null
-    }
-
-    var login: String? = null
-    if (index < bytes.size) {
-        val loginSize = bytes[index].toInt()
-        index += 1
-        if (loginSize > 0 && index + loginSize <= bytes.size) {
-            val loginBytes = bytes.copyOfRange(index, index + loginSize)
-            index += loginSize
-            login = String(loginBytes, Charsets.UTF_8)
-        }
-    }
-
-    var url: String? = null
-    if (index < bytes.size) {
-        val urlSize = bytes[index].toInt()
-        index += 1
-        if (urlSize > 0 && index + urlSize <= bytes.size) {
-            val urlBytes = bytes.copyOfRange(index, index + urlSize)
-            index += urlSize
-            url = String(urlBytes, Charsets.UTF_8)
-        }
-    }
-
-    return SecretData(
-        password = password,
-        login = login ?: "",
-        url = url ?: "",
-        label = "",
-        type = SeedkeeperSecretType.PASSWORD,
-        size = 0
-    )
-}
-
-fun countWords(mnemonic: String): Int {
-    return mnemonic.split("\\s+".toRegex()).filter { it.isNotEmpty() }.size
 }
 
 val instructionsMap: Map<Byte, String> = mapOf(
